+++ src/server_manager/model/cloud.ts
@@ -13,7 +13,85 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 */
+
 /** Enumeration of supported cloud providers. */
 export enum CloudProviderId {
   DigitalOcean = 'DigitalOcean',
 }
+
+/** DigitalOcean REST API wrapper */
+export interface DigitalOceanApi {
+  /** @see https://developers.digitalocean.com/documentation/v2/#get-user-information */
+  getAccount(): Promise<Account>;
+
+  /** @see */
+  createDroplet(
+      displayName: string, region: string, publicKeyForSSH: string,
+      dropletSpec: DigitalOceanDropletSpecification): Promise<{droplet: DropletInfo}>;
+
+  /** @see */
+  deleteDroplet(dropletId: number): Promise<void>;
+
+  /** @see https://developers.digitalocean.com/documentation/v2/#regions */
+  getRegionInfo(): Promise<RegionInfo[]>;
+
+  /** @see https://developers.digitalocean.com/documentation/v2/#retrieve-an-existing-droplet-by-id */
+  getDroplet(dropletId: number): Promise<DropletInfo>;
+
+  /** @see */
+  getDropletTags(dropletId: number): Promise<string[]>;
+
+  /** @see */
+  getDropletsByTag(tag: string): Promise<DropletInfo[]>;
+
+  /** @see */
+  getDroplets(): Promise<DropletInfo[]>;
+}
+
+export interface DigitalOceanDropletSpecification {
+  installCommand: string;
+  size: string;
+  image: string;
+  tags: string[];
+}
+
+export type DropletInfo = Readonly<{
+  id: number;
+  status: 'new' | 'active';
+  tags: string[];
+  region: {
+    readonly slug: string;
+  };
+  size: Readonly<{
+    transfer: number;
+    price_monthly: number;
+  }>;
+  networks: Readonly<{
+    v4: ReadonlyArray<
+        Readonly<{
+          type: string;
+          ip_address: string;}
+            >>;
+  }>;
+}>;
+
+export type Account = Readonly<{
+  email: string;
+  uuid: string;
+  email_verified: boolean;
+  status: string;
+}>;
+
+export type RegionInfo = Readonly<{
+  slug: string;
+  name: string;
+  sizes: string[];
+  available: boolean;
+  features: string[];
+}>;
+
+export type DigitalOceanError = Readonly<{
+  id: string;
+  message: string;
+  request_id?: string;
+}>;
